requirements.txt
streamlit
tensorflow
numpy
Pillow
torch
torchvision
ftfy
regex
timm
transformers
scikit-learn
git+https://github.com/openai/CLIP.git
